+++ formula-android/src/test/java/com/instacart/formula/android/fakes/FakeComponent.kt
@@ -1,6 +1,6 @@
 package com.instacart.formula.android.fakes
 
-import com.instacart.formula.fragment.FragmentContract
+import com.instacart.formula.android.FragmentContract
 import com.instacart.formula.android.FragmentKey
 import com.instacart.formula.android.DisposableScope
 import com.jakewharton.rxrelay3.PublishRelay